requirements.txt
pytube
openai
openai-whisper
pytesseract
langchain
faiss-cpu
transformers
torch
neo4j
pillow